deploy/requirements.txt
aws-cdk-lib>=2.110.0
constructs>=10.3.0
aws-cdk.aws-s3-deployment>=2.110.0
